+++ examples/resources/meraki_appliance_one_to_one_nat_rules/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_one_to_one_nat_rules.example "<network_id>"